requirements.txt
fastapi==0.99.0
uvicorn==0.22.0
sentence-transformers==2.2.2
torch>=2.0.0
cryptography>=41.0.0
httpx==0.24.1
numpy==1.26.0
pytest==7.4.0
python-dotenv==1.0.0
httpx
numpy
prometheus_client
streamlit



